requirements.txt
asgiref==3.9.2
async-timeout==5.0.1
certifi==2025.8.3
charset-normalizer==3.4.3
dj-database-url==3.0.1
Django==5.2.6
django-redis==6.0.0
djangorestframework==3.16.1
gunicorn==23.0.0
idna==3.10
packaging==25.0
psycopg2-binary==2.9.11
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.5
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
whitenoise==6.11.0
